[ALSA] add a private field for ac97-device drivers and let ucb1400 be its first user
From: Sebastian Siewior <bigeasy@linutronix.de>
Currently the UCB1400 driver discovers the interrupt via probing. This works
probably only on x86. This patch adds a private field to the ac97 struct
where the ac97 driver can deposit informations for the device driver that
serves a device which is attached to the ac97 bus.
This patch also converts the UCB1400 driver to use this information if
available.

Signed-off-by: Sebastian Siewior <bigeasy@linutronix.de>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>