ASoC Platform Driver for AT32AP7000 (AVR32)
Add ASoC platform driver for AT32AP7000 (AVR32).
This is basically a port of the at91 ASoC platform code to the
AVR32.  This is currently against the Linux 2.6.24.3.atmel.3 kernel,
the latest kernel Atmel has released for the AVR32.

playpaq_wm8510.c isn't necessarily intended for inclusion into the mainline
kernel or linux-2.6-asoc repository.  It's provided more as a
reference machine driver for an AT32AP7000 system.

Signed-off-by: Geoffrey Wossum <gwossum@acm.org>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>