Revised AT32 ASoC Patch
Attached is a revised version of my patch to add AT32 to ASoC.  This cleans up
most of the style issues associated with the previous patch.  Also fixes an
issue with the playpaq_wm8510.c code depending on a non-released patch to the
AT32 portmux support.

Patch is against 2.6.24.3.atmel.3 kernel, the latest AVR32 kernel Atmel has
released, with the linux-2.6-asoc patches from when v2.6.24 was tagged also applied.

Will apply cleanly and compile to a vanilla 2.6.24, with the following exceptions:

 - The sound/soc/Makefile patch is against the linux-2.6-asoc Makefile, not the
   vanilla Makefile
 - playpaq_wm8510.c requires the WM8510 driver to compile which is only
   currently available in linux-2.6-asoc

Signed-off-by: Geoffrey Wossum <gwossum@acm.org>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>