ASoC: Implement initial S3C64XX IISv4 support
The IISv4 port can be controlled by the IISv2 driver so copy over that
with light modifications, including use of platform data to get the pin
configuration. Currently this is S3C6410 specific due to the DMA
configuration - in the future when we have IISv4 specific support we may
want to split out a generic IISv4 portion.

Since this allows the SMDK64xx machine driver to work update it to
reflect the API changes involved and mark it as not BROKEN.

Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>